+++ pyscf/lib/np_helper/np_helper.h
@@ -49,9 +49,9 @@ void NPztranspose_021(int *shape, double complex *a, double complex *at);
 
 void NPomp_d_itranspose_scale(const int n, const double alpha, double *A, int lda);
 void NPomp_z_itranspose_scale(const int n, const double complex *alphaptr, double complex *A, int lda);
-void NPomp_dtensor_itranspose_scale021(const ssize_t matstride, int nmat, int n, const double alpha,
+void NPomp_dtensor_itranspose_scale021(const long long matstride, int nmat, int n, const double alpha,
                                       double *A, int lda);
-void NPomp_ztensor_itranspose_scale021(const ssize_t matstride, int nmat, int n, const double complex *alpha,
+void NPomp_ztensor_itranspose_scale021(const long long matstride, int nmat, int n, const double complex *alpha,
                                       double complex *A, int lda);
 
 void NPdunpack_tril_2d(int count, int n, double *tril, double *mat, int hermi);
